refactor: 해당 controller 일관성 있게 리팩토링
