+++ src/main/java/com/project/trainingdiary/controller/NotificationController.java
@@ -1,7 +1,7 @@
 package com.project.trainingdiary.controller;
 
-import com.project.trainingdiary.dto.request.RegisterFcmTokenRequestDto;
-import com.project.trainingdiary.dto.response.NotificationResponseDto;
+import com.project.trainingdiary.dto.request.notification.RegisterFcmTokenRequestDto;
+import com.project.trainingdiary.dto.response.notification.NotificationResponseDto;
 import com.project.trainingdiary.service.FcmTokenService;
 import com.project.trainingdiary.service.NotificationService;
 import io.swagger.v3.oas.annotations.Operation;
@@ -14,7 +14,6 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
-import org.springframework.data.domain.Sort.Direction;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PutMapping;
@@ -44,7 +43,8 @@ public class NotificationController {
       @RequestParam(defaultValue = "0") int page,
       @RequestParam(defaultValue = "20") int size
   ) {
-    Pageable pageable = PageRequest.of(page, size, Sort.by(Direction.DESC, "createdAt"));
+    Sort.Direction direction = Sort.Direction.DESC;
+    Pageable pageable = PageRequest.of(page, size, Sort.by(direction, "createdAt"));
     return ResponseEntity.ok(notificationService.getNotificationList(pageable));
   }
