--- src/main/java/com/project/trainingdiary/controller/TraineeController.java
@@ -1,5 +1,0 @@
-package com.project.trainingdiary.controller;
-
-public class TraineeController {
-
-}
